requirements.txt
certifi==2021.10.8
cycler==0.11.0
distlib==0.3.4
et-xmlfile==1.1.0
filelock==3.6.0
fonttools==4.33.3
kiwisolver==1.4.2
matplotlib==3.5.2
numpy==1.22.3
openpyxl==3.0.9
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
pipenv==2022.4.30
platformdirs==2.5.1
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
virtualenv==20.14.1
virtualenv-clone==0.5.7
xlrd==2.0.1
